Move from size_t to int for index_len
Silences a warning.  There was no benefit in using size_t, as index_len
is a realtivly small value.

+++ libiview/destroy_index.c
@@ -1,7 +1,7 @@
 #include <unistd.h>
 #include "iview.h"
 
-void iv_destroy_index(struct iv_series *index, size_t len) {
+void iv_destroy_index(struct iv_series *index, int len) {
     int i;
     for(i=0; i<len; i++) {
         free((void *)(index[i].title));
